Move functions that query models to models.py
